requirements.txt
telethon==1.26.0
asyncpg==0.25.0
python-dotenv==1.0.0
